modules/location_kafka_consumer/requirements.txt
kafka-python
SQLAlchemy==1.3.19
psycopg2-binary
pyrsistent
python-dateutil==2.8.1